Fixed response headers not being sent properly when more than one header of the same kind was set by the app.
An example of this is the cookie header. Try a simple app that sets two
cookies, only the first will reach the client.

Using Net::HTTP with such an app results in a malformed header's error.

Here's [a gist](https://gist.github.com/dariocravero/9102385) containing a test app
together with what could be the beginning of the test suite :).