fix potential race condition in async responses
Was possible for the response to arrive before the response processing
was complete, leading to data being thrown away. This is now resolved
by currying the request into the async callback and promises cannot be
resolved twice so the async callback can be called many times and only
the first response will be returned. This might however be a good
technique to allow HTTP2 to send multiple responses.